+++ src/tests/test_graphalgorithm.cc
@@ -62,7 +62,7 @@ BOOST_AUTO_TEST_CASE(single_network_algorithm_test) {
     GraphNode gn1;
     GraphNode gn2;
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     nodes[0] = gn1;
     nodes[1] = gn2;
 
@@ -101,7 +101,7 @@ BOOST_AUTO_TEST_CASE(single_network_algorithm_test) {
     GraphNode gn2;
     GraphNode gn3;
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     nodes[0] = gn1;
     nodes[1] = gn2;
     nodes[2] = gn3;
@@ -138,7 +138,7 @@ BOOST_AUTO_TEST_CASE(decouple_isolated_subgraphs_test) {
 
   vector<Edge> edges{ed1, ed2, ed3, ed4, ed5, ed6, ed7, ed8, ed9};
 
-  unordered_map<int, GraphNode> nodes;
+  unordered_map<long int, GraphNode> nodes;
   for (int index = 1; index < 12; ++index) {
     GraphNode gn;
     nodes[index] = gn;
@@ -167,7 +167,7 @@ BOOST_AUTO_TEST_CASE(decouple_isolated_subgraphs_test) {
   vertices_sub_graph3[11] = false;
 
   for (const Graph& sub_graph : sub_graphs) {
-    const vector<int> vertices = sub_graph.getVertices();
+    const vector<long int> vertices = sub_graph.getVertices();
     if (vertices_sub_graph1.count(vertices.at(0))) {
       for (const int& vertex : vertices) {
         if (vertices_sub_graph1.count(vertex)) {
@@ -241,7 +241,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
     GraphNode gn6;
     GraphNode gn7;
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     nodes[0] = gn1;
     nodes[1] = gn2;
     nodes[2] = gn3;
@@ -292,7 +292,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
     //
     vector<Edge> edges{Edge(1, 2), Edge(2, 3), Edge(3, 4), Edge(1, 4)};
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int vertex = 1; vertex <= 4; ++vertex) {
       GraphNode gn;
       nodes[vertex] = gn;
@@ -349,7 +349,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
     vector<Edge> edges{Edge(1, 2), Edge(2, 3), Edge(3, 4),
                        Edge(1, 4), Edge(3, 5), Edge(5, 6)};
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int vertex = 1; vertex <= 6; ++vertex) {
       GraphNode gn;
       nodes[vertex] = gn;
@@ -396,7 +396,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
                        Edge(3, 5), Edge(5, 7), Edge(7, 6), Edge(6, 8),
                        Edge(8, 9), Edge(9, 7)};
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int vertex = 1; vertex <= 9; ++vertex) {
       GraphNode gn;
       nodes[vertex] = gn;
@@ -460,7 +460,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
 
     vector<Edge> edges{ed1, ed2, ed3, ed4, ed5, ed6};
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int index = 0; index < 18; ++index) {
       GraphNode gn;
       nodes[index] = gn;
@@ -542,7 +542,7 @@ BOOST_AUTO_TEST_CASE(reduceGraph_algorithm_test) {
                        ed9, ed10, ed11, ed12, ed13, ed14, ed15, ed16};
 
     // Create Graph nodes
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int index = 0; index < 18; ++index) {
       GraphNode gn;
       nodes[index] = gn;
@@ -669,7 +669,7 @@ BOOST_AUTO_TEST_CASE(explorebranch_test) {
     vector<Edge> edges{ed1, ed2, ed3, ed4, ed5, ed6, ed7, ed8, ed9, ed10, ed11};
 
     // Create Graph nodes
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     for (int index = 0; index < 11; ++index) {
       GraphNode gn;
       nodes[index] = gn;
@@ -754,7 +754,7 @@ BOOST_AUTO_TEST_CASE(structureid_test) {
     GraphNode gn6;
     GraphNode gn7;
 
-    unordered_map<int, GraphNode> nodes;
+    unordered_map<long int, GraphNode> nodes;
     nodes[0] = gn1;
     nodes[1] = gn2;
     nodes[2] = gn3;
