+++ src/tests/test_reducededge.cc
@@ -44,21 +44,21 @@ BOOST_AUTO_TEST_CASE(equivalence_test) {
   BOOST_CHECK_EQUAL(ed, ed3);
   BOOST_CHECK_EQUAL((ed == ed2), false);
 
-  ReducedEdge ed4(vector<int>{2, 3, 4});
-  ReducedEdge ed5(vector<int>{4, 3, 2});
+  ReducedEdge ed4(vector<long int>{2, 3, 4});
+  ReducedEdge ed5(vector<long int>{4, 3, 2});
   BOOST_CHECK_EQUAL(ed4, ed5);
   BOOST_CHECK_EQUAL((ed4 == ed), false);
 
-  ReducedEdge ed6(vector<int>{2, 6, 3, 1});
-  ReducedEdge ed7(vector<int>{1, 3, 6, 2});
+  ReducedEdge ed6(vector<long int>{2, 6, 3, 1});
+  ReducedEdge ed7(vector<long int>{1, 3, 6, 2});
 
   auto chain = ed6.getChain();
   BOOST_CHECK_EQUAL(ed6, ed7);
 
   // Both are loops they get rearranged internally so the smallest vertex
   // appears as EndPoint1 and EndPoint2
-  ReducedEdge ed8(vector<int>{1, 2, 4, 1});
-  ReducedEdge ed9(vector<int>{2, 4, 1, 2});
+  ReducedEdge ed8(vector<long int>{1, 2, 4, 1});
+  ReducedEdge ed9(vector<long int>{2, 4, 1, 2});
   BOOST_CHECK_EQUAL(ed8, ed9);
 }
 
@@ -77,23 +77,23 @@ BOOST_AUTO_TEST_CASE(getter_test) {
   BOOST_CHECK_EQUAL(ed.getEndPoint1(), 2);
   BOOST_CHECK_EQUAL(ed.getEndPoint2(), 3);
 
-  ReducedEdge ed3(vector<int>{2, 6, 3, 1});
+  ReducedEdge ed3(vector<long int>{2, 6, 3, 1});
   BOOST_CHECK_EQUAL(ed3.getEndPoint1(), 1);
   BOOST_CHECK_EQUAL(ed3.getEndPoint2(), 2);
-  ReducedEdge ed4(vector<int>{1, 3, 6, 2});
+  ReducedEdge ed4(vector<long int>{1, 3, 6, 2});
   BOOST_CHECK_EQUAL(ed4.getEndPoint1(), 1);
   BOOST_CHECK_EQUAL(ed4.getEndPoint2(), 2);
 
-  ReducedEdge ed5(vector<int>{1, 2, 4, 1});
+  ReducedEdge ed5(vector<long int>{1, 2, 4, 1});
   BOOST_CHECK_EQUAL(ed5.getEndPoint1(), 1);
   BOOST_CHECK_EQUAL(ed5.getEndPoint2(), 1);
-  ReducedEdge ed6(vector<int>{2, 4, 1, 2});
+  ReducedEdge ed6(vector<long int>{2, 4, 1, 2});
   BOOST_CHECK_EQUAL(ed6.getEndPoint1(), 1);
   BOOST_CHECK_EQUAL(ed6.getEndPoint2(), 1);
 }
 
 BOOST_AUTO_TEST_CASE(getchain_test) {
-  vector<int> vertices{4, 1, 6, 7, 0};
+  vector<long int> vertices{4, 1, 6, 7, 0};
   ReducedEdge ed(vertices);
   auto chain = ed.getChain();
   BOOST_CHECK_EQUAL(chain.at(0), 0);
@@ -104,7 +104,7 @@ BOOST_AUTO_TEST_CASE(getchain_test) {
 }
 
 BOOST_AUTO_TEST_CASE(vertexwithinchain_test) {
-  vector<int> vertices{1, 3, 4, 5};
+  vector<long int> vertices{1, 3, 4, 5};
   ReducedEdge ed(vertices);
   BOOST_CHECK(ed.vertexExistInChain(1));
   BOOST_CHECK(ed.vertexExistInChain(2) == false);
@@ -124,20 +124,20 @@ BOOST_AUTO_TEST_CASE(less_test) {
   ReducedEdge ed5(2, 2);
   BOOST_CHECK_EQUAL(ed4 < ed5, true);
 
-  ReducedEdge ed6(vector<int>{1, 2, 3});
+  ReducedEdge ed6(vector<long int>{1, 2, 3});
   BOOST_CHECK_EQUAL(ed1 < ed6, true);
 
-  ReducedEdge ed7(vector<int>{1, 2, 3, 4, 1});
-  ReducedEdge ed8(vector<int>{1, 4, 3, 2, 1});
+  ReducedEdge ed7(vector<long int>{1, 2, 3, 4, 1});
+  ReducedEdge ed8(vector<long int>{1, 4, 3, 2, 1});
   BOOST_CHECK_EQUAL((ed7 < ed8), false);  // Equal
 
-  ReducedEdge ed9(vector<int>{1, 4, 5, 2, 1});
+  ReducedEdge ed9(vector<long int>{1, 4, 5, 2, 1});
   BOOST_CHECK_EQUAL((ed7 < ed9), true);
 
-  ReducedEdge ed10(vector<int>{1, 4, 5, 0, 1});
+  ReducedEdge ed10(vector<long int>{1, 4, 5, 0, 1});
   BOOST_CHECK_EQUAL((ed10 < ed9), true);
 
-  ReducedEdge ed11(vector<int>{1, 0, 5, 2, 1});
+  ReducedEdge ed11(vector<long int>{1, 0, 5, 2, 1});
   BOOST_CHECK_EQUAL((ed11 < ed9), true);
 }
 
@@ -153,7 +153,7 @@ BOOST_AUTO_TEST_CASE(hash_key_test) {
   set<ReducedEdge> e_set;
   ReducedEdge ed(23, 43);
   e_set.insert(ed);
-  unordered_map<int, ReducedEdge> e_map;
+  unordered_map<long int, ReducedEdge> e_map;
   e_map[2] = ed;
 }
